DDD UI support

+++ OpenKh.Tools.ModsManager/ViewModels/MainViewModel.cs
@@ -455,6 +455,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                     ConfigOpenKhGameEngineLocation = ConfigurationService.OpenKhGameEngineLocation,
                     ConfigPcsx2Location = ConfigurationService.Pcsx2Location,
                     ConfigPcReleaseLocation = ConfigurationService.PcReleaseLocation,
+                    ConfigPcReleaseLocationKH3D = ConfigurationService.PcReleaseLocationKH3D,
                     ConfigPcReleaseLanguage = ConfigurationService.PcReleaseLanguage,
                     ConfigRegionId = ConfigurationService.RegionId,
                     ConfigPanaceaInstalled = ConfigurationService.PanaceaInstalled,
@@ -468,6 +469,7 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                     ConfigurationService.OpenKhGameEngineLocation = dialog.ConfigOpenKhGameEngineLocation;
                     ConfigurationService.Pcsx2Location = dialog.ConfigPcsx2Location;
                     ConfigurationService.PcReleaseLocation = dialog.ConfigPcReleaseLocation;
+                    ConfigurationService.PcReleaseLocationKH3D = dialog.ConfigPcReleaseLocationKH3D;
                     ConfigurationService.RegionId = dialog.ConfigRegionId;
                     ConfigurationService.PanaceaInstalled = dialog.ConfigPanaceaInstalled;
                     ConfigurationService.IsEGSVersion = dialog.ConfigIsEGSVersion;
@@ -478,6 +480,16 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                         Directory.Exists(ConfigurationService.PcReleaseLocation))
                     {
                         File.WriteAllLines(Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt"),
+                            new string[]
+                            {
+                                $"mod_path={ConfigurationService.GameModPath}",
+                                $"show_console={false}",
+                            });
+                    }
+                    if (ConfigurationService.GameEdition == EpicGamesPC &&
+                        Directory.Exists(ConfigurationService.PcReleaseLocationKH3D))
+                    {
+                        File.WriteAllLines(Path.Combine(ConfigurationService.PcReleaseLocationKH3D, "panacea_settings.txt"),
                             new string[]
                             {
                                 $"mod_path={ConfigurationService.GameModPath}",
@@ -596,52 +608,74 @@ namespace OpenKh.Tools.ModsManager.ViewModels
                 case 2:
                     if (ConfigurationService.IsEGSVersion && !(_launchGame == "kh3d"))
                     {
-                        if (ConfigurationService.PanaceaInstalled)
+                        if (ConfigurationService.PcReleaseLocation != null)
                         {
-                            string panaceaSettings = Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt");
-                            if (!File.Exists(panaceaSettings))
+                            if (ConfigurationService.PanaceaInstalled)
                             {
-                                File.WriteAllLines(Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt"),
-                                new string[]
+                                string panaceaSettings = Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt");
+                                if (!File.Exists(panaceaSettings))
                                 {
+                                    File.WriteAllLines(Path.Combine(ConfigurationService.PcReleaseLocation, "panacea_settings.txt"),
+                                    new string[]
+                                    {
                                 $"mod_path={ConfigurationService.GameModPath}",
                                 $"show_console={false}",
-                                });
+                                    });
+                                }
+                                File.AppendAllText(panaceaSettings, "\nquick_launch=" + _launchGame);
                             }
-                            File.AppendAllText(panaceaSettings, "\nquick_launch=" + _launchGame);
+                            processStartInfo = new ProcessStartInfo
+                            {
+                                FileName = "com.epicgames.launcher://apps/4158b699dd70447a981fee752d970a3e%3A5aac304f0e8948268ddfd404334dbdc7%3A68c214c58f694ae88c2dab6f209b43e4?action=launch&silent=true",
+                                UseShellExecute = true,
+                            };
                         }
-                        processStartInfo = new ProcessStartInfo
+                        else
                         {
-                            FileName = "com.epicgames.launcher://apps/4158b699dd70447a981fee752d970a3e%3A5aac304f0e8948268ddfd404334dbdc7%3A68c214c58f694ae88c2dab6f209b43e4?action=launch&silent=true",
-                            UseShellExecute = true,
-                        };
+                            MessageBox.Show(
+                           "Unable to locate KINGDOM HEARTS HD 1.5+2.5 ReMIX install. Please re-run the setup wizard and confirm it is correct.",
+                           "Run error", MessageBoxButton.OK, MessageBoxImage.Error);
+                            CloseAllWindows();
+                            return Task.CompletedTask;
+                        }
                     }
                     else if (ConfigurationService.IsEGSVersion && _launchGame == "kh3d")
                     {
-                        string panaceaSettings = Path.Combine(ConfigurationService.PcReleaseLocationKH3D, "panacea_settings.txt");
-                        if (ConfigurationService.PanaceaInstalled)
+                        if (ConfigurationService.PcReleaseLocationKH3D != null)
                         {
-                            if (!File.Exists(panaceaSettings))
+                            string panaceaSettings = Path.Combine(ConfigurationService.PcReleaseLocationKH3D, "panacea_settings.txt");
+                            if (ConfigurationService.PanaceaInstalled)
                             {
-                                if (Directory.Exists(ConfigurationService.PcReleaseLocationKH3D))
+                                if (!File.Exists(panaceaSettings))
                                 {
-                                    File.WriteAllLines(Path.Combine(ConfigurationService.PcReleaseLocationKH3D, "panacea_settings.txt"),
-                                    new string[]
+                                    if (Directory.Exists(ConfigurationService.PcReleaseLocationKH3D))
                                     {
+                                        File.WriteAllLines(Path.Combine(ConfigurationService.PcReleaseLocationKH3D, "panacea_settings.txt"),
+                                        new string[]
+                                        {
                                     $"mod_path={ConfigurationService.GameModPath}",
                                     $"show_console={false}",
-                                    });
+                                        });
+                                    }
+                                    else
+                                        File.Create(panaceaSettings);
                                 }
-                                else
-                                    File.Create(panaceaSettings);
+                                File.AppendAllText(panaceaSettings, "\nquick_launch=" + _launchGame);
                             }
-                            File.AppendAllText(panaceaSettings, "\nquick_launch=" + _launchGame);
+                            processStartInfo = new ProcessStartInfo
+                            {
+                                FileName = "com.epicgames.launcher://apps/c8ff067c1c984cd7ab1998e8a9afc8b6%3Aaa743b9f52e84930b0ba1b701951e927%3Ad1a8f7c478d4439b8c60a5808715dc05?action=launch&silent=true",
+                                UseShellExecute = true,
+                            };
                         }
-                        processStartInfo = new ProcessStartInfo
+                        else
                         {
-                            FileName = "com.epicgames.launcher://apps/c8ff067c1c984cd7ab1998e8a9afc8b6%3Aaa743b9f52e84930b0ba1b701951e927%3Ad1a8f7c478d4439b8c60a5808715dc05?action=launch&silent=true",
-                            UseShellExecute = true,
-                        };
+                            MessageBox.Show(
+                           "Unable to locate KINGDOM HEARTS HD 2.8 Final Chapter Prologue install. Please re-run the setup wizard and confirm it is correct.",
+                           "Run error", MessageBoxButton.OK, MessageBoxImage.Error);
+                            CloseAllWindows();
+                            return Task.CompletedTask;
+                        }
                     }
                     else
                     {
